backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
pandas>=2.0,<3.0
numpy>=1.26,<3.0
scipy>=1.11,<2.0
scikit-learn>=1.4,<2.0
PyMySQL>=1.1,<2.0
matplotlib>=3.8,<4.0
seaborn>=0.13,<0.14
cryptography>=42.0.0,<45.0.0
